Route for admin login, different secret for admin tokens
