+++ .env
@@ -4,3 +4,4 @@ POSTGRES_DB=rvdb
 PORT=8081
 NODE_ENV=development
 JWT_SECRET=devsecret
+JWT_ADMIN_SECRET=admindevsecret
