+++ src/app.js
@@ -13,10 +13,14 @@
     var account_route = require('./routes/account');
     var register_route = require('./routes/register');
     var purchase_route = require('./routes/purchase');
+    var admin_auth = require('./routes/admin/adminAuth');
+
     app.use('/api/v1/user/authenticate', auth_route);
     app.use('/api/v1/user/account', account_route);
     app.use('/api/v1/user/register', register_route);
     app.use('/api/v1/product/purchase', purchase_route);
 
+    app.use('/api/v1/admin/authenticate', admin_auth);
+    
     module.exports = app;
 }());
