+++ src/jwt/token.js
@@ -1,12 +1,11 @@
 const jwt = require('jsonwebtoken');
-const tokenSecret = process.env.JWT_SECRET;
 const expiration = Math.floor(Date.now() / 1000) + 86400;
 
-module.exports.sign = function (payload) {
+module.exports.sign = function (payload, tokenSecret = process.env.JWT_SECRET) {
     return jwt.sign({ exp: expiration, data: payload }, tokenSecret, { algorithm: 'HS256' });
 };
 
-module.exports.verify = function (jwtToken) {
+module.exports.verify = function (jwtToken, tokenSecret = process.env.JWT_SECRET) {
     var decoded = null;
 
     try {
